Fake v1 version for v2 framework for now
The package manager uses `WP_CLI_VERSION` to require the framework. Through the `VERSION` file, the framework now has the version `2.0.0-alpha` already. This causes package installs to fail, as all packages require `v1.x.x` framework as a dependency.

This PR intercepts v2.x versions of the framework in the package manager and fakes them to be versioned at `1.99.99-alpha`, so that it still works with existing packages.

This is a temporary measure. The correct solution is to adapt the packages so that they either work with both v1 & v2, or are known and accpeted to not work with v2 due to breaking changes.

Related https://github.com/wp-cli/scaffold-package-command/issues/167